create upsertListItems and move it into list-utils
